GroupG: relocate 12 test files to correct sub-packages and extract 2 handler tests
PM-01: tests/test_analyze_prs_gates.py → tests/pipeline/test_pr_gates.py
PM-02: tests/test_phase2_cook_session.py → tests/server/test_headless_session.py
PM-03: tests/test_phase2_headless_env.py → tests/execution/test_headless_env_injection.py
PM-04: tests/test_phase2_hooks.py → tests/infra/test_open_kitchen_guard.py
PM-05: tests/test_phase2_session_skills.py → tests/workspace/test_session_skills.py
PM-06: tests/test_pr_domain_partitioner.py → tests/pipeline/test_pr_domain_partitioner.py
PM-07: tests/test_pyproject_metadata.py → tests/infra/test_pyproject_metadata.py (+path fix)
PM-08: tests/test_release_sanity.py → tests/infra/test_release_sanity.py (+path fix)
PM-09: tests/test_review_pr_fidelity.py → tests/pipeline/test_fidelity.py
PM-10: tests/pipeline/test_branch_guard.py → tests/core/test_branch_guard.py
PM-11: tests/recipe/test_mcp_overrides.py → tests/server/test_mcp_overrides.py
PM-12: tests/server/test_helpers.py → tests/config/test_helpers.py
PM-13: extract 2 server handler tests from execution/test_ci_params.py into
        server/test_tools_ci.py; remove now-unused json/AsyncMock imports
