+++ tests/infra/test_open_kitchen_guard.py
