+++ tests/server/test_headless_session.py
