+++ tests/workspace/test_session_skills.py
